Drop ST3 support
