+++ LSP-metals.sublime-commands
@@ -7,15 +7,15 @@
     { "caption": "LSP-metals: Compile Cancel", "command": "lsp_metals_execute", "args":{"command_name": "compile-cancel"}},
     { "caption": "LSP-metals: Doctor Run", "command": "lsp_metals_execute", "args":{"command_name": "doctor-run"}},
     { "caption": "LSP-metals: Reset Choice", "command": "lsp_metals_execute", "args":{"command_name": "reset-choice"}},
-    { "caption": "LSP-metals: Show Decoded TASTy", "command": "file_decoder", "args":{"decoding_type": "tasty-decoded"}},
-    { "caption": "LSP-metals: Show decompiled with javap", "command": "file_decoder", "args":{"decoding_type": "javap"}},
-    { "caption": "LSP-metals: Show decompiled with javap verbose", "command": "file_decoder", "args":{"decoding_type": "javap-verbose"}},
-    { "caption": "LSP-metals: Show semanticdb", "command": "file_decoder", "args":{"decoding_type": "semanticdb-compact"}},
-    { "caption": "LSP-metals: Show semanticdb detailed", "command": "file_decoder", "args":{"decoding_type": "semanticdb-detailed"}},
-    { "caption": "LSP-metals: Show semanticdb proto", "command": "file_decoder", "args":{"decoding_type": "semanticdb-proto"}},
-    { "caption": "LSP-metals: Find in dependency", "command": "find_in_dependency", "args":{}},
-    { "caption": "LSP-metals: New Scala File", "command": "new_scala_file", "args":{"paths": []}},
-    { "caption": "LSP-metals: Analyze Stacktrace", "command": "analyze_stacktrace"},
+    { "caption": "LSP-metals: Show Decoded TASTy", "command": "lsp_file_decoder", "args":{"decoding_type": "tasty-decoded"}},
+    { "caption": "LSP-metals: Show decompiled with javap", "command": "lsp_metals_file_decoder", "args":{"decoding_type": "javap"}},
+    { "caption": "LSP-metals: Show decompiled with javap verbose", "command": "lsp_metals_file_decoder", "args":{"decoding_type": "javap-verbose"}},
+    { "caption": "LSP-metals: Show semanticdb", "command": "lsp_metals_file_decoder", "args":{"decoding_type": "semanticdb-compact"}},
+    { "caption": "LSP-metals: Show semanticdb detailed", "command": "lsp_metals_file_decoder", "args":{"decoding_type": "semanticdb-detailed"}},
+    { "caption": "LSP-metals: Show semanticdb proto", "command": "lsp_metals_file_decoder", "args":{"decoding_type": "semanticdb-proto"}},
+    { "caption": "LSP-metals: Find in dependency", "command": "lsp_metals_find_in_dependency", "args":{}},
+    { "caption": "LSP-metals: New Scala File", "command": "lsp_metals_new_scala_file", "args":{"paths": []}},
+    { "caption": "LSP-metals: Analyze Stacktrace", "command": "lsp_metals_analyze_stacktrace"},
     {
         "caption": "LSP-metals: Settings",
         "command": "edit_settings",
