+++ commands/lsp_metals_text_command.py
@@ -1,0 +1,5 @@
+from LSP.plugin.core.registry import LspTextCommand
+
+
+class LspMetalsTextCommand(LspTextCommand):
+    session_name = "metals"
